refactor: rm duplicated until() function
