+++ packages/client/test/end-to-end/StorageNodeRegistry.test.ts
@@ -1,7 +1,8 @@
 import 'reflect-metadata'
 import { StreamrClient } from '../../src/StreamrClient'
 import { Wallet } from 'ethers'
-import { createTestStream, fetchPrivateKeyWithGas, until } from '../test-utils/utils'
+import { createTestStream, fetchPrivateKeyWithGas } from '../test-utils/utils'
+import { until } from '../../src/utils'
 import { Stream } from '../../src'
 import { ConfigTest, DOCKER_DEV_STORAGE_NODE } from '../../src/ConfigTest'
 import { afterAll } from 'jest-circus'
