+++ packages/client/test/test-utils/utils.ts
@@ -539,48 +539,6 @@ export function getWaitForStorage(client: StreamrClient, defaultOpts = {}): (las
     }
 }
 
-/**
- * Wait until a condition is true
- * @param condition - wait until this callback function returns true
- * @param timeOutMs - stop waiting after that many milliseconds, -1 for disable
- * @param pollingIntervalMs - check condition between so many milliseconds
- * @param failedMsgFn - append the string return value of this getter function to the error message, if given
- * @return the (last) truthy value returned by the condition function
- */
-export async function until(condition: MaybeAsync<() => boolean>, timeOutMs = 10000, pollingIntervalMs = 100, failedMsgFn?: () => string): Promise<boolean> {
-    // condition could as well return any instead of boolean, could be convenient
-    // sometimes if waiting until a value is returned. Maybe change if such use
-    // case emerges.
-    const err = new Error(`Timeout after ${timeOutMs} milliseconds`)
-    let isTimedOut = false
-    let t!: ReturnType<typeof setTimeout>
-    if (timeOutMs > 0) {
-        t = setTimeout(() => { isTimedOut = true }, timeOutMs)
-    }
-
-    try {
-        // Promise wrapped condition function works for normal functions just the same as Promises
-        let wasDone = false
-        while (!wasDone && !isTimedOut) { // eslint-disable-line no-await-in-loop
-            wasDone = await Promise.resolve().then(condition) // eslint-disable-line no-await-in-loop
-            if (!wasDone && !isTimedOut) {
-                await wait(pollingIntervalMs) // eslint-disable-line no-await-in-loop
-            }
-        }
-
-        if (isTimedOut) {
-            if (failedMsgFn) {
-                err.message += ` ${failedMsgFn()}`
-            }
-            throw err
-        }
-
-        return wasDone
-    } finally {
-        clearTimeout(t)
-    }
-}
-
 export const createEthereumAddress = (id: number): string => {
     return '0x' + padEnd(String(id), 40, '0')
 }
